Adding Map Test and wireData mechanism for Skandium

+++ src/tests/bpskel/shared/tasks/DataPrintTask.java
@@ -8,7 +8,7 @@ public class DataPrintTask extends AbstractSimpleTask {
 
 	@Override
 	public void execute() throws ExecutionException {
-		System.out.println((int)this.getResultData().getData());
+		System.out.println(this.getResultData().getData());
 	}
 
 }
